Added copyright header and moved files around  #14184
